correction in utils parse double

+++ src/gui/util/Utils.java
@@ -31,7 +31,7 @@ public class Utils {
 	
 	public static Double tryParseToDouble(String str) {
 		try {
-			return Double.parseInt(str);
+			return Double.parseDouble(str);
 		} catch (NumberFormatException e) {
 			return null;
 		}
